\name{con_gene_name}
\alias{con_gene_name}
\docType{data}
\title{
gene name/char ...
}
\description{
 control gene name of the dataset. 
}
\usage{data(con_gene_name)}
\format{
  The format is:
 chr "18s"
}

\examples{
data(con_gene_name)
## maybe str(con_gene_name) ; plot(con_gene_name) ...
}
\keyword{datasets}
