\name{con_group_name}
\alias{con_group_name}
\docType{data}
\title{
 group name/char 
}
\description{
 control group of the dataset.
}
\usage{data(con_group_name)}
\format{
  The format is:
  char "CON"
}

\references{
 ~~ possibly secondary sources and usages ~~
}
\examples{
data(con_group_name)
## maybe str(con_group_name) ; plot(con_group_name) ...
}
\keyword{datasets}
