\name{filename}
\alias{filename}
\docType{data}
\title{
 filename/character 
}
\description{
  ~~ Specify the path and file name of the csv file ~~
}
\usage{data(filename)}
\format{
  The format is:
 char "c:/qpcr/rawdata.csv"
}

\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(filename)
## maybe str(filename) ; plot(filename) ...
}
\keyword{datasets}
