\name{qPCR.CT-package}
\alias{qPCR.CT-package}
\alias{qPCR.CT}
\docType{package}
\title{
qPCR data analysis and plot package
}
\description{
use 2^ddCT methods calculate the relative gene expression, data file can be export from bio-rad qpcr machine, the results can be plot with errorbar.
ver 1.1 add GroupPlot function, can plot all the groups once.
}
\details{
\tabular{ll}{
Package: \tab qPCR.CT\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2011-09-12\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
Depends: \tab RColorBrewer \cr
}
}
\author{
Yuzhuo Pan,  Xiaoyu Yan, Junxin Li

Maintainer: Yuzhuo Pan <yuzhuo.pan@gmail.com>
}
\keyword{ qPCR, Plot, RColorBrewer }
\seealso{
RColorBrewer
}
